Ajust import dirs in DynFlags
We need to adjust the import paths passed to GHC via the "-i" flag.
This is tricky to get right when working with the raw command-line
arguments to GHC, because of the existence of flags like
"-instantiated-with" and "-ignore-package". With this change, we
make this adjustment after GHC parses its command-line arguments,
by modify the relevant field of the 'DynFlags'.

The adjustment to the import paths needs the directory which those
paths are relative to. I have added a field to 'ComponentOptions'
to propogagte that directory to where we call GHC.